Reorganize all TopMeta components into subdirectory
