+++ packages/frontend/amp/components/topMeta/PaidForBand.tsx
