+++ packages/frontend/amp/components/topMeta/TopMeta.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { StyledTone } from '@frontend/amp/lib/tag-utils';
 import { ArticleModel } from '@frontend/amp/pages/Article';
-import { TopMetaNews } from '@frontend/amp/components/TopMetaNews';
-import { TopMetaOpinion } from '@frontend/amp/components/TopMetaOpinion';
-import { TopMetaPaidContent } from '@frontend/amp/components/TopMetaPaidContent';
+import { TopMetaNews } from '@frontend/amp/components/topMeta/TopMetaNews';
+import { TopMetaOpinion } from '@frontend/amp/components/topMeta/TopMetaOpinion';
+import { TopMetaPaidContent } from '@frontend/amp/components/topMeta/TopMetaPaidContent';
 
 export const TopMeta: React.SFC<{ data: ArticleModel; tone: StyledTone }> = ({
     data,
